The captain files the tests where the crew can inspect them (library boundary and Clippy)
The standards review had left the game as a binary-only crate with its behavioral suites embedded under production modules, while strict Clippy still rejected several API shapes and UI calls.\n\nAdd a shared src/lib.rs surface, move all 212 behavioral cases into documented integration suites under 	ests/, and keep the native entry point as a thin consumer. Replace argument-heavy risk, ledger, and rendering calls with named request/view structs, resolve the remaining Clippy diagnostics, activate the empty source-size gate, and document the project workflow in README.md.\n\nThe existing catalog metadata change is staged with the project working tree as required by the repository checklist. cargo fmt -- --check, cargo test --all-targets, and cargo clippy --all-targets -- -D warnings pass.

+++ src/state/port.rs
@@ -30,8 +30,7 @@ impl GameSession {
         }
         self.ship_layout.remove(module_id);
         self.economy.credits -= module.remove_cost;
-        self.career
-            .record_module_change(i64::from(module.remove_cost));
+        self.career.record_module_change(module.remove_cost);
         self.damaged_modules.retain(|id| id != module_id);
         Ok(format!(
             "Removed {} for {} credits",
@@ -81,8 +80,7 @@ impl GameSession {
             .place(module_id, module.footprint, position, rotation, true)
             .map_err(|error| error.to_string())?;
         self.economy.credits -= module.purchase_cost;
-        self.career
-            .record_module_change(i64::from(module.purchase_cost));
+        self.career.record_module_change(module.purchase_cost);
         if !self.unlocked_modules.iter().any(|id| id == module_id) {
             self.unlocked_modules.push(module_id.to_owned());
         }
